+++ lib/frames/contents/ignored_frame_content.dart
@@ -2,20 +2,24 @@ import 'package:zooper_flutter_id3/frames/headers/id3v2_frame_header.dart';
 import 'package:zooper_flutter_id3/frames/models/raw_model.dart';
 import 'package:zooper_flutter_id3/tags/headers/id3_header.dart';
 
-import 'id3v2_frame_content.dart';
+import 'frame_content.dart';
 
-class IgnoredFrameContent extends Id3v2FrameContent<RawModel> {
-  IgnoredFrameContent.decode(
+class IgnoredFrameContent extends FrameContent<RawModel> {
+  factory IgnoredFrameContent.decode(
     Id3Header header,
     Id3v2FrameHeader frameHeader,
     List<int> bytes,
     int startIndex,
     int size,
-  ) : super() {
+  ) {
     var content = bytes.sublist(startIndex, startIndex + size);
 
-    model = RawModel(content);
+    var model = RawModel(content);
+
+    return IgnoredFrameContent(model);
   }
+
+  IgnoredFrameContent(RawModel model) : super(model);
 
   @override
   List<int> encode() {