+++ lib/frames/id3v2_frame.dart
@@ -1,14 +1,15 @@
-import 'package:zooper_flutter_id3/frames/contents/id3v2_frame_content.dart';
+import 'package:zooper_flutter_id3/frames/contents/frame_content.dart';
+import 'package:zooper_flutter_id3/frames/contents/id3v2_frame_content_factory.dart';
 import 'package:zooper_flutter_id3/frames/id3_frame.dart';
 import 'package:zooper_flutter_id3/frames/headers/id3v2_frame_header.dart';
 import 'package:zooper_flutter_id3/tags/headers/id3_header.dart';
 import 'package:zooper_flutter_id3/tags/headers/id3v2_header.dart';
 
-class Id3v2Frame extends Id3Frame<Id3v2FrameHeader, Id3v2FrameContent> {
+class Id3v2Frame extends Id3Frame<Id3v2FrameHeader, FrameContent> {
   const Id3v2Frame(
     Id3Header header,
     Id3v2FrameHeader frameHeader,
-    Id3v2FrameContent frameContent,
+    FrameContent frameContent,
   ) : super(frameHeader, frameContent);
 
   static Id3v2Frame? decode(
@@ -26,7 +27,7 @@ class Id3v2Frame extends Id3Frame<Id3v2FrameHeader, Id3v2FrameContent> {
     if (frameHeader == null) return null;
 
     // Decode the content
-    var frameContent = Id3v2FrameContent.decode(
+    var frameContent = Id3v2FrameContentFactory.decode(
       header,
       frameHeader,
       bytes,