+++ lib/tags/contents/id3v1_content.dart
@@ -5,6 +5,13 @@ import 'package:zooper_flutter_id3/frames/id3v1_frame.dart';
 import 'package:zooper_flutter_id3/tags/contents/id3_content.dart';
 import 'package:zooper_flutter_id3/tags/headers/id3_header.dart';
 
+import '../../exceptions/id3_exception.dart';
+import '../../frames/contents/frame_content.dart';
+import '../../frames/headers/frame_header.dart';
+import '../../frames/id3_frame.dart';
+import '../../frames/models/frame_content_model.dart';
+import '../../frames/models/text_model.dart';
+
 class Id3v1Content extends Id3Content<Id3v1Frame> {
   Id3v1Content.decode(Id3Header header, List<int> bytes, int startIndex) {
     // Title
@@ -76,6 +83,34 @@ class Id3v1Content extends Id3Content<Id3v1Frame> {
         frame.frameHeader.identifier.frameName == FrameName.year ||
         frame.frameHeader.identifier.frameName == FrameName.comment ||
         frame.frameHeader.identifier.frameName == FrameName.contentType;
+  }
+
+  @override
+  void deleteFrame(Id3Frame<FrameHeader, FrameContent<FrameContentModel>> frame) {
+    deleteFramesByName(frame.frameHeader.identifier.frameName);
+  }
+
+  @override
+  void deleteFramesByName(FrameName frameName) {
+    var countFrames = countFramesByIdentifier(frameName);
+
+    if (countFrames == 0) {
+      return;
+    }
+
+    if (countFrames > 1) {
+      // Something is wrong here
+      throw Id3Exception('ID3v1 Frame $frameName exists multiple times');
+    }
+
+    var model = getContentModelsByFrameName(frameName).first;
+
+    /// Because every Id3v1 Frame is existent, we don't delete a frame, we set it's content to 'Nothing'
+    if (model is Id3v1TextModel) {
+      model.value = '';
+    } else {
+      throw UnimplementedError('Frame Model ${model.runtimeType} is not implemented yet');
+    }
   }
 
   int _loadFrame<T>(Id3Header header, FrameIdentifier identifier, List<int> bytes, int startIndex) {