+++ lib/tags/id3_tag.dart
@@ -1,6 +1,7 @@
-import 'package:collection/collection.dart';
+import 'package:quiver/collection.dart';
 import 'package:zooper_flutter_id3/enums/frame_name.dart';
 import 'package:zooper_flutter_id3/frames/id3_frame.dart';
+import 'package:zooper_flutter_id3/frames/models/frame_content_model.dart';
 import 'package:zooper_flutter_id3/tags/contents/id3_content.dart';
 
 import 'headers/id3_header.dart';
@@ -20,9 +21,22 @@ abstract class Id3Tag<THeader extends Id3Header, TContent extends Id3Content, TF
   }
 
   /// Returns the Frame stored inside, or null
-  TFrame? getFrameByNameOrNull(FrameName name) {
-    return content.frames.firstWhereOrNull((element) => element.frameHeader.identifier.frameName == name) as TFrame?;
-  }
+  List<TFrame> getFramesByName(FrameName name) => content.getFramesByName(name).cast<TFrame>().toList();
+
+  /// Gets all available frames
+  List<TFrame> getFrames() => content.frames.cast<TFrame>();
+
+  /// Returns the Frames aas a Multimap
+  ///
+  /// A multimap is a type found inside
+  /// [https://pub.dev/documentation/quiver/latest/quiver.collection/Multimap-class.html]
+  Multimap<FrameName, FrameContentModel> getFramesAsMultimap() => content.getFramesAsMultimap();
+
+  /// Returns all the models of a given frame
+  ///
+  /// Note: Some frames can only exist once, some can be present multiple times
+  List<FrameContentModel> getContentModelsByFrameName(FrameName frameName) =>
+      content.getContentModelsByFrameName(frameName);
 
   /// Checks if a frame already exists by it's name
   ///
@@ -30,10 +44,14 @@ abstract class Id3Tag<THeader extends Id3Header, TContent extends Id3Content, TF
   /// which are not defined as "the same" frame, this will not check if the
   /// frame is one of this frames and has a different description.
   /// So if there is such a frame, this method will always return TRUE
-  bool containsFrameWithIdentifier(FrameName name) {
-    return content.frames.any((element) => element.frameHeader.identifier.frameName == name);
-  }
+  bool containsFrameWithIdentifier(FrameName frameName) => content.containsFrameWithIdentifier(frameName);
 
   /// Encodes the tag
   List<int> encode();
+
+  /// Deletes a specific frame
+  void deleteFrame(Id3Frame frame) => content.deleteFrame(frame);
+
+  /// Deletes all frames with the given identifier
+  void deleteFramesByName(FrameName frameName) => content.deleteFramesByName(frameName);
 }