+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 0.0.1-alpha
 
 Basic reading and writing of Frames
+
+## 0.0.2-alpha
+
+- Added ability to delete frames
+- Added ability to access frame models easily
+- Added ability to read the attached picture